created context to pass data as props, modularized MatchedDog, FavoriteDogs, SearchInput
